requirements.txt
Django==4.2.7
psycopg2-binary>=2.9.9
Pillow>=10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
